refactor: utils folder refactored
