+++ client/src/lib/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { RootState } from 'store/modules/combine-reducer';
-import { displayErrorMessage } from 'utils';
+import { displayErrorMessage } from 'lib/utils';
 
 const ErrorBoundary: React.FC = () => {
   const { error } = useSelector((state: RootState) => state.error);
